Clean up more front end warnings.
Refactor GMT class slightly.

Add a README for dummy directory.
