+++ anf/bin/web/usarray_deploy_map/anf/Makefile
@@ -1,1 +1,5 @@
+# Dummy makefile so that the child directories get included.
+# As this level is just for namespacing during development, we have no files to install here.
+
+#!!! Please do not add anything to be installed here directly !!!
 include $(ANFMAKE)
