+++ anf/bin/web/usarray_deploy_map/anf/deploymentmap/exceptions.py
@@ -2,7 +2,7 @@
 
 from anf.logutil import getLogger
 
-logger = getLogger(__name__)
+LOGGER = getLogger(__name__)
 
 
 class DeployMapError(Exception):
